Type fixes

+++ ErrorEst/NoiseUnit.h
@@ -5,15 +5,12 @@
 
 #include <vector>
 #include <cmath>
-//long double Mo (const long double *x,int l);
-long double Mo (std::vector<long double> const &x,int l);
-long double Sko (const long double *x0,const long double *x,int l);
-//long double Sko (std::vector<long double> const &x0,std::vector<long double> const &x,int l);
-//void ShumAdding(std::vector<long double> const *x,long double *out,long double *ret, long double koeff,const int l);
-void ShumAdding(std::vector<long double> const &x,std::vector<long double> &out,std::vector<long double> &ret, long double koeff,const int l);
-
-void QuantumShumAdding(std::vector<long double> const &x,std::vector<long double> &out,
-    std::vector<long double> &ret,const int l,long double mzr, int amplitude);
+long double Mo (std::vector<long double> const &x, unsigned long l);
+long double Sko (const long double *x0,const long double *x,unsigned long l);
+void ShumAdding(std::vector<long double> const &x, std::vector<long double> &out,
+                std::vector<long double> &ret, long double koeff, const unsigned long l);
+void QuantumShumAdding(std::vector<long double> const &x, std::vector<long double> &out,
+    std::vector<long double> &ret, const unsigned long l, long double mzr, int amplitude);
 
 //---------------------------------------------------------------------------
 #endif
